requirements.txt
streamlit
pandas
scikit-learn
sentence-transformers
faiss-cpu
openai
azure-applicationinsights
python-dotenv
